fix: clean the config
